Update patch set 3
Patch Set 3:

So this was a namespace resolution problem. The MediaWikiServices class is in the `MediaWiki\MediaWikiServices` namespace. Before this patch, the resolution couldn't happen as PHP could see that the services instance is being created in a wrong namespace (which should fail?)

Now this patch corrects it and I suggest using `use` to import the namespace and a shorter line when creating the services instance the file :) Thanks @Paladox.

Patch-set: 3
CC: D3r1ck01 <2842@e9e9afe9-4712-486d-8885-f54b72dd1951>